Small fixes after SCP-3509: add runtime checks for allowed builtins
